+++ src/Data/Serialize/Describe.hs
@@ -1,0 +1,7 @@
+module Data.Serialize.Describe(
+  module Data.Serialize.Describe.Descriptor,
+  module Data.Serialize.Describe.Class
+) where
+
+import Data.Serialize.Describe.Descriptor
+import Data.Serialize.Describe.Class